Refactor Meta and MetaZMQ into its own files. Create fig_to_src() in utils and move sanitize_message() there too.
