+++ databench/__init__.py
@@ -5,9 +5,12 @@ from __future__ import absolute_import
 
 __version__ = "0.4a19"
 
-from .analysis import Meta, Analysis, sanitize_message
-from .analysis_zmq import MetaZMQ, AnalysisZMQ
+from .analysis import Analysis
+from .analysis_zmq import AnalysisZMQ
 from .app import App
 from .datastore import Datastore
+from .meta import Meta
+from .meta_zmq import MetaZMQ
 from .readme import Readme
 from .testing import AnalysisTestCase, AnalysisTestCaseSSL
+from .utils import sanitize_message, fig_to_src
